pyproject.toml
[project] # Project metadata
name = "thorvg-python"
description = "A ctypes API for thorvg, a vector graphics engine supporting SVG and Lottie formats."
requires-python = ">=3.8"
readme = "README.md"
license = "LGPL-2.1-only"
authors = [{ "name" = "chaudominic", "email" = "chaudominic2@gmail.com" }]
keywords = ["thorvg", "lottie", "svg", "ctypes", "python"]
classifiers = ["Topic :: Multimedia :: Graphics"]
dependencies = []
version = "1.0.0"

[project.urls]
Repository = "https://github.com/laggykiller/thorvg-python"
Documentation = "https://thorvg-python.readthedocs.io/en/latest/"
Tracker = "https://github.com/laggykiller/thorvg-python/issues"

[project.optional-dependencies]
full = [
    "Pillow",
    "numpy",
]

test = [
    "pytest",
]

lint = [
    "ruff",
    "mypy",
    "isort",
    "types-Pillow",
]

[build-system]
requires = [
    "conan>=2.0",
    "setuptools>=45",
    "lipomerge>=0.1.1"
]
build-backend = "setuptools.build_meta"

[tool.pyright]
include = ["src", "tests"]
strict = ["*"]

[tool.mypy]
python_version = "3.9"
files = ["src", "tests"]